Convert entrypoints to TS
