--- src/entrypoint_factory.js
@@ -1,21 +1,0 @@
-'use strict';
-
-const m = require('mithril');
-const FeatureFlagService = require('App/services/feature_flag');
-
-const { SignInForm } = require('App/views/common/auth');
-const { App, Welcome } = require('App/views/factory');
-const { FactorySignUpForm, FactoryDetails } = require('App/views/factory/details');
-const AuthService = require('App/services/auth');
-const { ListCertifications } = require('App/views/factory/requests');
-
-AuthService.namespace = 'factory';
-
-const element = document.getElementById('app');
-m.route(element, '/', {
-    '/': App.subpage(Welcome),
-    '/signUp': App.subpage(FeatureFlagService.isSignupEnabled() && FactorySignUpForm),
-    '/signIn': App.subpage(SignInForm),
-    '/profile': App.subpage(FactoryDetails),
-    '/requests': App.subpage(ListCertifications),
-});
