+++ src/entrypoint_retailer.ts
@@ -1,13 +1,10 @@
-'use strict';
-
-const m = require('mithril');
-
-const { App, Welcome } = require('App/views/retailer');
-const { AgentList } = require('App/views/retailer/agent');
-const { OrganizationCreate, OrganizationList } = require('App/views/common/organization');
-const { Certifications } = require('App/views/retailer/search');
-const { FactoryList } = require('App/views/auditor/factory');
-const { FactoryProfile } = require('App/views/retailer/factory_profile');
+import * as m from 'mithril';
+import { App, Welcome } from 'App/views/retailer';
+import { AgentList } from 'App/views/retailer/agent';
+import { OrganizationCreate, OrganizationList } from 'App/views/common/organization';
+import { Certifications } from 'App/views/retailer/search';
+import { FactoryList } from 'App/views/auditor/factory';
+import { FactoryProfile } from 'App/views/retailer/factory_profile';
 
 const element = document.getElementById('app');
 m.route(element, '/', {
